chg: dev: Refactor helpers
