+++ lib/redmine_reveal.rb
@@ -5,8 +5,7 @@ require_relative 'redmine_reveal/patches/controller_patch'
 require_relative 'redmine_reveal/patches/wiki_content_patch'
 
 # Helpers
-require_relative 'redmine_reveal/helpers/textile_helper'
-require_relative 'redmine_reveal/helpers/markdown_helper'
+require_relative 'redmine_reveal/helpers/wiki_formatting'
 
 # Hooks
 require_relative 'redmine_reveal/hooks/view_hooks'
